Iterate through rotated file names rather than files in the directory
This allows more flexible file rotation naming schemes, using date
formats, so the "extension" attribute was added, using an extended
form of strftime.

The "start_num" attribute was added, allowing the starting number to
be changed.

The "if_missing" attribute was added, which specifies whether files
will be rotated if the "file" is missing.  By default, it is false.